:sparkles: vue loading
