+++ src/main/java/com/tale/model/params/ThemeParam.java
@@ -1,0 +1,14 @@
+package com.tale.model.params;
+
+import lombok.Data;
+
+/**
+ * @author biezhi
+ * @date 2018/6/10
+ */
+@Data
+public class ThemeParam {
+
+    private String siteTheme;
+
+}
