+++ src/main/resources/static/admin/js/edit_article.js
@@ -23,7 +23,11 @@ var vm = new Vue({
             createdTime: moment().format('YYYY-MM-DD HH:mm:ss'),
             selected: ['默认分类']
         },
-        categories: []
+        categories: [],
+        isLoading: true
+    },
+    beforeCreate: function(){
+        vueLoding = this.$loading.show();
     },
     mounted: function () {
         var $vm = this;
@@ -353,5 +357,8 @@ $(document).ready(function () {
             });
         }
     });
+
+    vm.isLoading = false;
+    vueLoding.hide();
 
 });
