+++ src/main/resources/templates/admin/article/_footer.html
@@ -1,12 +1,12 @@
 #include('../footer.html')
 <script src="${cdnURL()}/plugins/tagsinput/jquery.tagsinput.min.js"></script>
-<script async src="${cdnURL()}/plugins/datetimepicker/bootstrap-datetimepicker.min.js"></script>
-<script async src="${cdnURL()}/plugins/datetimepicker/bootstrap-datetimepicker.zh-CN.js"></script>
-<script async src="${cdnURL()}/plugins/jquery-multi-select/jquery.quicksearch.js"></script>
+<script src="${cdnURL()}/plugins/datetimepicker/bootstrap-datetimepicker.min.js"></script>
+<script src="${cdnURL()}/plugins/datetimepicker/bootstrap-datetimepicker.zh-CN.js"></script>
+<script src="${cdnURL()}/plugins/jquery-multi-select/jquery.quicksearch.js"></script>
 <script src="${cdnURL()}/plugins/mditor/js/mditor.min.js"></script>
-<script async src="${cdnURL()}/plugins/wysihtml5/0.3.0/wysihtml5.min.js"></script>
-<script async src="${cdnURL()}/plugins/summernote/0.8.2/summernote.min.js"></script>
-<script async src="${cdnURL()}/plugins/summernote/0.8.2/lang/summernote-zh-CN.min.js"></script>
+<script src="${cdnURL()}/plugins/wysihtml5/0.3.0/wysihtml5.min.js"></script>
+<script src="${cdnURL()}/plugins/summernote/0.8.2/summernote.min.js"></script>
+<script src="${cdnURL()}/plugins/summernote/0.8.2/lang/summernote-zh-CN.min.js"></script>
 <script src="${cdnURL()}/plugins/multi-select/0.9.12/js/jquery.multi-select.min.js"></script>
 <script src="${cdnURL()}/plugins/select2/3.4.8/select2.min.js"></script>
 <script src="${cdnURL()}/plugins/jquery-toggles/2.0.4/toggles.min.js"></script>
