+++ src/main/resources/templates/comm/ts_base_layout.html
@@ -3,7 +3,6 @@
 ${bodyContent}
 </div>
 #include('/admin/footer.html')
-#include('/comm/tale_comment.html')
 <script type="text/javascript">
 
     var tale = new $.tale();
