+++ src/main/resources/templates/themes/default/setting.html
@@ -84,4 +84,4 @@
     </div>
 </form>
 #end
-#include ("/comm/ts_base_layout.html")
+#include("/comm/ts_base_layout.html")
